Moved functions into functions folder. Added "get slack profile" microservice and referenced from other relevant lambdas. Added office and cohort fields to cert table, and added logic to update it in the deactivateFins lambda, now renamed to updateFins.
