+++ buildspec.yml
@@ -32,10 +32,10 @@ phases:
 
       # For environment var SLACK_APP_AUTH_TOKEN in serverless.yml (OAuth Access Token https://api.slack.com/apps/A9ZA995GF/oauth?success=1)
       - aws s3api get-object --bucket $(cat s3.txt) --key oauth.enc oauth.txt
+      # For environment var SLACK_APP_USER_AUTH_TOKEN in serverless.yml (OAuth Access Token https://api.slack.com/apps/A9ZA995GF/oauth?success=1)
+      - aws s3api get-object --bucket $(cat s3.txt) --key oauth_user.enc oauth_user.txt
       # For environment var VERIFICATION_TOKEN in serverless.yml
       - aws s3api get-object --bucket $(cat s3.txt) --key token.enc token.txt
-      # For environment var SLACK_BOT_AUTH_TOKEN in serverless.yml
-      #- aws s3api get-object --bucket $(cat s3.txt) --key bot_oauth.enc bot_oauth.txt
 
       # For environment var EPSAGON_TOKEN in serverless.yml
       - aws s3api get-object --bucket $(cat s3.txt) --key epsagon.enc epsagon.txt
@@ -45,7 +45,7 @@ phases:
   build:
     commands:
       # Deploy Application
-      - serverless deploy --clientSecret "$(cat oauth.txt)" --slackToken "$(cat token.txt)" --epsagonToken "$(cat epsagon.txt)"
+      - serverless deploy --clientSecret "$(cat oauth.txt)" --slackToken "$(cat token.txt)" --epsagonToken "$(cat epsagon.txt)" --clientSecret_user "$(cat oauth_user.txt)"
 
   post_build:
     commands:
